storage: do some refactoring
* move sqlite specific code to new package
* use shutdown context for all queries
* move db file migration code to sqlite3 pkg
