+++ pkg/storage/sqlite/certificates_get.go
@@ -39,7 +39,7 @@ func (store *Storage) GetAllCerts(q pagination_sort.Query) (certs []certificates
 	sort := sortField + " " + q.SortDirection()
 
 	// do query
-	ctx, cancel := context.WithTimeout(context.Background(), store.timeout)
+	ctx, cancel := context.WithTimeout(store.shutdownContext, store.timeout)
 	defer cancel()
 
 	// WARNING: SQL Injection is possible if the variables are not properly
@@ -193,7 +193,7 @@ func (store *Storage) GetOneCertByName(name string) (cert certificates.Certifica
 
 // getOneCert returns a Cert based on either its unique id or its unique name
 func (store *Storage) getOneCert(id int, name string) (cert certificates.Certificate, err error) {
-	ctx, cancel := context.WithTimeout(context.Background(), store.timeout)
+	ctx, cancel := context.WithTimeout(store.shutdownContext, store.timeout)
 	defer cancel()
 
 	query := `
