+++ pkg/storage/sqlite/keys_post.go
@@ -8,7 +8,7 @@ import (
 // PostNewKey saves the KeyExtended to the db as a new key
 func (store *Storage) PostNewKey(payload private_keys.NewPayload) (private_keys.Key, error) {
 	// database action
-	ctx, cancel := context.WithTimeout(context.Background(), store.timeout)
+	ctx, cancel := context.WithTimeout(store.shutdownContext, store.timeout)
 	defer cancel()
 
 	query := `
