+++ pkg/storage/sqlite/setup_migrate_v1.go
@@ -258,9 +258,7 @@ func (store *Storage) migrateV0toV1() (int, error) {
 	oldSchemaVer := 0
 	newSchemaVer := 1
 
-	store.logger.Infof("updating database user_version from %d to %d", oldSchemaVer, newSchemaVer)
-
-	ctx, cancel := context.WithTimeout(context.Background(), store.timeout)
+	ctx, cancel := context.WithTimeout(store.shutdownContext, store.timeout)
 	defer cancel()
 
 	// create sql transaction to roll back in the event an error occurs
@@ -361,6 +359,5 @@ func (store *Storage) migrateV0toV1() (int, error) {
 		return -1, err
 	}
 
-	store.logger.Infof("database user_version successfully upgraded from %d to %d", oldSchemaVer, newSchemaVer)
 	return newSchemaVer, nil
 }
