+++ pkg/storage/sqlite/setup_migrate_v10.go
@@ -17,9 +17,7 @@ func (store *Storage) migrateV9toV10() (int, error) {
 	oldSchemaVer := 9
 	newSchemaVer := 10
 
-	store.logger.Infof("updating database user_version from %d to %d", oldSchemaVer, newSchemaVer)
-
-	ctx, cancel := context.WithTimeout(context.Background(), store.timeout)
+	ctx, cancel := context.WithTimeout(store.shutdownContext, store.timeout)
 	defer cancel()
 
 	// create sql transaction to roll back in the event an error occurs
@@ -79,6 +77,5 @@ func (store *Storage) migrateV9toV10() (int, error) {
 		return -1, err
 	}
 
-	store.logger.Infof("database user_version successfully upgraded from %d to %d", oldSchemaVer, newSchemaVer)
 	return newSchemaVer, nil
 }
