+++ pkg/storage/sqlite/setup_migrate_v11.go
@@ -184,9 +184,7 @@ func (store *Storage) migrateV10toV11() (int, error) {
 	oldSchemaVer := 10
 	newSchemaVer := 11
 
-	store.logger.Infof("updating database user_version from %d to %d", oldSchemaVer, newSchemaVer)
-
-	ctx, cancel := context.WithTimeout(context.Background(), store.timeout)
+	ctx, cancel := context.WithTimeout(store.shutdownContext, store.timeout)
 	defer cancel()
 
 	// create sql transaction to roll back in the event an error occurs
@@ -236,6 +234,5 @@ func (store *Storage) migrateV10toV11() (int, error) {
 		return -1, err
 	}
 
-	store.logger.Infof("database user_version successfully upgraded from %d to %d", oldSchemaVer, newSchemaVer)
 	return newSchemaVer, nil
 }
