+++ pkg/storage/sqlite/setup_migrate_v3.go
@@ -20,9 +20,7 @@ func (store *Storage) migrateV2toV3() (int, error) {
 	oldSchemaVer := 2
 	newSchemaVer := 3
 
-	store.logger.Infof("updating database user_version from %d to %d", oldSchemaVer, newSchemaVer)
-
-	ctx, cancel := context.WithTimeout(context.Background(), store.timeout)
+	ctx, cancel := context.WithTimeout(store.shutdownContext, store.timeout)
 	defer cancel()
 
 	// create sql transaction to roll back in the event an error occurs
@@ -89,6 +87,5 @@ func (store *Storage) migrateV2toV3() (int, error) {
 		return -1, err
 	}
 
-	store.logger.Infof("database user_version successfully upgraded from %d to %d", oldSchemaVer, newSchemaVer)
 	return newSchemaVer, nil
 }
