+++ pkg/storage/sqlite/setup_migrate_v4.go
@@ -15,9 +15,7 @@ func (store *Storage) migrateV3toV4() (int, error) {
 	oldSchemaVer := 3
 	newSchemaVer := 4
 
-	store.logger.Infof("updating database user_version from %d to %d", oldSchemaVer, newSchemaVer)
-
-	ctx, cancel := context.WithTimeout(context.Background(), store.timeout)
+	ctx, cancel := context.WithTimeout(store.shutdownContext, store.timeout)
 	defer cancel()
 
 	// create sql transaction to roll back in the event an error occurs
@@ -67,6 +65,5 @@ func (store *Storage) migrateV3toV4() (int, error) {
 		return -1, err
 	}
 
-	store.logger.Infof("database user_version successfully upgraded from %d to %d", oldSchemaVer, newSchemaVer)
 	return newSchemaVer, nil
 }
