+++ pkg/storage/sqlite/setup_migrate_v6.go
@@ -16,9 +16,7 @@ func (store *Storage) migrateV5toV6() (int, error) {
 	oldSchemaVer := 5
 	newSchemaVer := 6
 
-	store.logger.Infof("updating database user_version from %d to %d", oldSchemaVer, newSchemaVer)
-
-	ctx, cancel := context.WithTimeout(context.Background(), store.timeout)
+	ctx, cancel := context.WithTimeout(store.shutdownContext, store.timeout)
 	defer cancel()
 
 	// create sql transaction to roll back in the event an error occurs
@@ -69,6 +67,5 @@ func (store *Storage) migrateV5toV6() (int, error) {
 		return -1, err
 	}
 
-	store.logger.Infof("database user_version successfully upgraded from %d to %d", oldSchemaVer, newSchemaVer)
 	return newSchemaVer, nil
 }
