+++ pkg/storage/sqlite/setup_migrate_v7.go
@@ -17,9 +17,7 @@ func (store *Storage) migrateV6toV7() (int, error) {
 	oldSchemaVer := 6
 	newSchemaVer := 7
 
-	store.logger.Infof("updating database user_version from %d to %d", oldSchemaVer, newSchemaVer)
-
-	ctx, cancel := context.WithTimeout(context.Background(), store.timeout)
+	ctx, cancel := context.WithTimeout(store.shutdownContext, store.timeout)
 	defer cancel()
 
 	// create sql transaction to roll back in the event an error occurs
@@ -78,6 +76,5 @@ func (store *Storage) migrateV6toV7() (int, error) {
 		return -1, err
 	}
 
-	store.logger.Infof("database user_version successfully upgraded from %d to %d", oldSchemaVer, newSchemaVer)
 	return newSchemaVer, nil
 }
