+++ pkg/storage/sqlite/setup_migrate_v9.go
@@ -15,9 +15,7 @@ func (store *Storage) migrateV8toV9() (int, error) {
 	oldSchemaVer := 8
 	newSchemaVer := 9
 
-	store.logger.Infof("updating database user_version from %d to %d", oldSchemaVer, newSchemaVer)
-
-	ctx, cancel := context.WithTimeout(context.Background(), store.timeout)
+	ctx, cancel := context.WithTimeout(store.shutdownContext, store.timeout)
 	defer cancel()
 
 	// create sql transaction to roll back in the event an error occurs
@@ -80,6 +78,5 @@ func (store *Storage) migrateV8toV9() (int, error) {
 		return -1, err
 	}
 
-	store.logger.Infof("database user_version successfully upgraded from %d to %d", oldSchemaVer, newSchemaVer)
 	return newSchemaVer, nil
 }
